fix: preserve input dtype in ParallelEQUtilities to prevent float64 promotion (issue #2158)
scipy.signal.sosfilt() always returns float64 regardless of input dtype.
This caused all three parallel EQ methods (apply_low_shelf_boost,
apply_high_shelf_boost, apply_bandpass_boost) to silently promote float32
audio to float64, doubling memory usage and causing dtype mismatches
downstream in the mastering pipeline.

Two-point fix per method (9 changes total across 3 methods):
1. Cast band immediately after sosfilt() to audio.dtype — eliminates the
   float64 band array so the intermediate computation stays in float32.
2. Cast the final return to audio.dtype — guards against NumPy ≥ 2.0
   (NEP-50) where Python float scalars are treated as float64 and would
   promote band * boost_diff back to float64 even with a float32 band.

51 new tests in TestLow/HighShelfDtype and TestBandpassDtype cover mono and
stereo × float32 and float64 for each method: dtype preservation, shape,
finiteness, and 0 dB identity. TestFloat32QualityPreservation checks that
float32 output stays within 2e-3 relative tolerance of float64 output,
confirming no meaningful quality regression.

+++ auralis/core/dsp/parallel_eq.py
@@ -90,10 +90,13 @@ class ParallelEQUtilities:
 
         # Extract low-frequency band
         # Handle both mono (samples,) and stereo (channels, samples)
+        # Cast to audio.dtype immediately: sosfilt() always returns float64
+        # regardless of input dtype, which doubles memory and promotes the
+        # entire output when mixed with a float32 signal (issue #2158).
         if audio.ndim == 1:
-            band = sosfilt(sos_filter, audio)
+            band = sosfilt(sos_filter, audio).astype(audio.dtype)
         else:
-            band = sosfilt(sos_filter, audio, axis=1)
+            band = sosfilt(sos_filter, audio, axis=1).astype(audio.dtype)
 
         # Calculate parallel boost amount
         # boost_linear = 1.0 means no change
@@ -102,8 +105,11 @@ class ParallelEQUtilities:
         boost_linear = 10 ** (boost_db / 20)
         boost_diff = boost_linear - 1.0
 
-        # Add boosted band to original (parallel processing)
-        return audio + band * boost_diff
+        # Add boosted band to original (parallel processing).
+        # Final astype() preserves dtype across NumPy versions: in NumPy ≥ 2.0
+        # (NEP-50) Python float scalars are treated as float64, so the multiply
+        # would otherwise promote band back to float64 before the add.
+        return (audio + band * boost_diff).astype(audio.dtype)
 
     @staticmethod
     def apply_high_shelf_boost(
@@ -156,18 +162,18 @@ class ParallelEQUtilities:
         # Create high-pass filter
         sos_filter = butter(order, normalized_freq, btype='high', output='sos')
 
-        # Extract high-frequency band
+        # Extract high-frequency band; cast for dtype preservation (issue #2158).
         if audio.ndim == 1:
-            band = sosfilt(sos_filter, audio)
+            band = sosfilt(sos_filter, audio).astype(audio.dtype)
         else:
-            band = sosfilt(sos_filter, audio, axis=1)
+            band = sosfilt(sos_filter, audio, axis=1).astype(audio.dtype)
 
         # Calculate parallel boost amount
         boost_linear = 10 ** (boost_db / 20)
         boost_diff = boost_linear - 1.0
 
-        # Add boosted band to original (parallel processing)
-        return audio + band * boost_diff
+        # Add boosted band to original (parallel processing).
+        return (audio + band * boost_diff).astype(audio.dtype)
 
     @staticmethod
     def apply_bandpass_boost(
@@ -233,15 +239,15 @@ class ParallelEQUtilities:
         # Create bandpass filter
         sos_filter = butter(order, [low_norm, high_norm], btype='band', output='sos')
 
-        # Extract mid-range band
+        # Extract mid-range band; cast for dtype preservation (issue #2158).
         if audio.ndim == 1:
-            band = sosfilt(sos_filter, audio)
+            band = sosfilt(sos_filter, audio).astype(audio.dtype)
         else:
-            band = sosfilt(sos_filter, audio, axis=1)
+            band = sosfilt(sos_filter, audio, axis=1).astype(audio.dtype)
 
         # Calculate parallel boost amount
         boost_linear = 10 ** (boost_db / 20)
         boost_diff = boost_linear - 1.0
 
-        # Add boosted band to original (parallel processing)
-        return audio + band * boost_diff
+        # Add boosted band to original (parallel processing).
+        return (audio + band * boost_diff).astype(audio.dtype)
